# Back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API en Laravel",
        "description": "Documentación de la API con Swagger en Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/canchas2": {
            "get": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Obtener todas las canchas",
                "operationId": "48d206ad7daadcff95ae95b55a8a2acd",
                "responses": {
                    "200": {
                        "description": "Lista de canchas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Crear una nueva cancha (Solo ADMIN)",
                "operationId": "3963bbbd4a28f4dfbbfcd03f3963dad8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "tipo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Cancha 1"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Fútbol 5"
                                    },
                                    "precio_hora": {
                                        "type": "number",
                                        "example": 1200
                                    },
                                    "cant_jugadores": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cancha creada correctamente"
                    },
                    "401": {
                        "description": "No autenticado (Falta Token)"
                    },
                    "403": {
                        "description": "Acceso denegado (Falta Rol)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/canchas2/{id}": {
            "get": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Obtener una cancha por ID",
                "operationId": "78f77dde1f489a905240d45a50078de6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos de la cancha"
                    },
                    "404": {
                        "description": "Cancha no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Actualizar una cancha existente (Solo ADMIN)",
                "operationId": "fa1ab91a1ad1b1ba2b7498c09e41e902",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Cancha 1"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Fútbol 7"
                                    },
                                    "precio_hora": {
                                        "type": "number",
                                        "example": 1500
                                    },
                                    "cant_jugadores": {
                                        "type": "integer",
                                        "example": 14
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cancha actualizada correctamente"
                    },
                    "401": {
                        "description": "No autenticado (Falta Token)"
                    },
                    "403": {
                        "description": "Acceso denegado (Falta Rol)"
                    },
                    "404": {
                        "description": "Cancha no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Canchas"
                ],
                "summary": "Eliminar una cancha (Solo ADMIN)",
                "operationId": "ef98a39ac76a0a209090fd0b5f6ee787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancha eliminada correctamente"
                    },
                    "401": {
                        "description": "No autenticado (Falta Token)"
                    },
                    "403": {
                        "description": "Acceso denegado (Falta Rol)"
                    },
                    "404": {
                        "description": "Cancha no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario existente (Solo Administrador)",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    },
                    "403": {
                        "description": "Acceso denegado (Requiere Administrador)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario (Solo Administrador)",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    },
                    "403": {
                        "description": "Acceso denegado (Requiere Administrador)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/usuarios-registrados": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener lista de usuarios registrados (Solo Administrador)",
                "operationId": "8bff2980165735f8f8aebb1f4e1d0792",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios"
                    },
                    "403": {
                        "description": "Acceso denegado (Requiere Administrador)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token JWT para autenticar",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}